refactor to a separate library
This commit refactors the existing code into a separate library that is
called by the main facter ruby file. This allows for this library to be
potentially used in other contexts.

Additionally, this refactor moves everything to be data structure
focused. Users of the fact can determine which parts of the fact they
wish to have added to facter.