add framework files for sections 2-6
With a specific file structure and structured input, it is possible to
auto-generate most of the fact creation ruby files. This commit has
framework files for sections 2 through 6 (CentOS 7), as well as the Perl
script that generated the files.

Note: These are only frameworked files, they have everything except the
full failure case test code. The framework is in place for this code, as
well as hints that were auto-generated.