handle multiple same executable audit steps
Before this commit, multiple steps using the same executable in the
audit process were being squashed down to only show the final step.
This commit fixes this by having all steps with the same executable show
with each subsequent step after the initial being tagged with a numeric
indicator. The engineer working on the output will then need to handle
it appropriately.

This commit also adds the (Scored)/(Not Scored) identifier to the test
description comment.